pyproject.toml
[project]
name = "api-burgers"
version = "0.1.0"
description = ""
authors = [
    {name = "advisor5",email = "advisor5@host.local"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pytest (>=8.4.1,<9.0.0)",
    "requests (>=2.32.4,<3.0.0)",
    "allure-pytest (>=2.15.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
